Added some ignore files with patterns to ignore eclipse stuff and class files
git-svn-id: http://svn.psas.pdx.edu/svn/psas/trunk/java@2204 6115d4fc-9e01-0410-8836-dad5f2c52705

+++ src/widgets/RocketState.java
@@ -107,7 +107,7 @@ public class RocketState extends JPanel implements Observer
 				setZ(msg.getData16(2));
 				break;
 			case CanBusIDs.FC_REPORT_STATE:
-				setState((int) msg.getData8(0) & 0xff);
+				setState(msg.getData8(0) & 0xff);
 				break;
 			case CanBusIDs.FC_REPORT_NODE_STATUS:
 				detailDisplay.setStates(msg.getBody());
